+++ aoc-utils/aoc-utils-kotlin/src/main/kotlin/de/havox_design/aoc/utils/kotlin/helpers/BooleanHelper.kt
@@ -1,0 +1,6 @@
+package de.havox_design.aoc.utils.kotlin.helpers
+
+fun Boolean.toInt() = when {
+    this -> 1
+    else -> 0
+}