+++ CHANGELOG.md
@@ -11,6 +11,7 @@ day 4.
 * Import `Iterable<A>.pairwise(other: Iterable<B>)` from 
 [AoC 2025](https://github.com/Gentleman1983/advent-of-code-2025) day 9.
 * Import `Iterable<A>.combinations()` from [AoC 2025](https://github.com/Gentleman1983/advent-of-code-2025) day 9.
+* Import `Boolean.toInt()` from [AoC 2025](https://github.com/Gentleman1983/advent-of-code-2025) day 10.
 
 ## 2025.2.0
 *  Update dependency org.sonarqube to v7.2.2.6593